Share structure_props and resolve_path across vscode webview
Extract the duplicated settings->Structure-props mapper and the JSON path
resolver into detect.ts; main.ts, JsonBrowser.svelte, and detect.test.ts
import the shared versions instead of keeping local copies.
